updating plan-card.scss
using hover transform for __view--front to rotate it 180 degrees, adding transform 180 degrees to __view--back, and then using hover transform for __view--back to rotate it 0 degrees. changing position of __view to absolute and top and left 0 so the front and back would be on top of each other. giving plan-card position relative and 50 rem for height. changing backface-visibility for __view to hidden.